// src/app/shared/model/api/courseDetailing.model.ts
export class CourseDetailing {
    apiId: number = 0;
    level: string = '';
    degree: string = '';
    modality: string = '';
    city: string = '';
    offerType: string = '';
    knowledgeArea: string = '';
    // knowledgeAxis: string | null = null;
    courseLoad: string | null = null;
    minimumCourseLoad: string | null = null;
    turn: string | null = null;
    courseSlots: string | null = null;
    apiName: string = '';
    apiNameFiltered: string = '';
    pnpName: string | null = null;
}